requirements.txt
# Bindings for the hermes protocol 
hermes-python>=0.1.29
grove.py>=0.6

# More dependency goes here.. 
pyserial==3.2.1
board
adafruit_dotstar